api_server/requirements.txt
flask
logging
threading
pyTelegramBotAPI
python-dotenv
